test: adds test for inactivity email job
