+++ config/env.yml
@@ -29,6 +29,7 @@ default: &default
     email_sender: "ana@ribon.io"
     name_sender: "Ana da Ribon"
     default_category: "All Contacts"
+    one_week_inactivity_template_id: "d-176422cd49e24a1b81fdb16054125cb3"
   web3:
     wallets:
       ribon_wallet_private_key: <%= ENV['RIBON_WALLET_PRIVATE_KEY'] %>
@@ -59,6 +60,12 @@ test:
     ribon_secret_openssl_key: ""
   polygon_scan:
     api_key: "YourApiKey"
+  mailer:
+    api_key: "sendgrid_api_key"
+    email_sender: "ana@ribon.io"
+    name_sender: "Ana da Ribon"
+    default_category: "All Contacts"
+    one_week_inactivity_template_id: "d-176422cd49e24a1b81fdb16054125cb3"
 
 production:
   <<: *default
